Refactor `SplineLink` into `Spline.Link`
