+++ app/src/commonMain/kotlin/diy/lingerie/geometry/splines/Spline.kt
@@ -1,13 +1,45 @@
 package diy.lingerie.geometry.splines
 
+import diy.lingerie.algebra.NumericObject
+import diy.lingerie.geometry.Point
 import diy.lingerie.geometry.curves.PrimitiveCurve
+import diy.lingerie.geometry.transformations.Transformation
 
 /**
  * A composite curve, at least positionally-continuous (C0), either open or
  * closed.
  */
 interface Spline {
-    val links: List<SplineLink>
+    data class Link(
+        val start: Point,
+        val edge: PrimitiveCurve.Edge,
+    ) : NumericObject {
+        fun bind(
+            end: Point,
+        ): PrimitiveCurve = edge.bind(
+            start = start,
+            end = end,
+        )
+
+        override fun equalsWithTolerance(
+            other: NumericObject,
+            tolerance: NumericObject.Tolerance,
+        ): Boolean = when {
+            other !is Link -> false
+            !start.equalsWithTolerance(other.start, tolerance) -> false
+            !edge.equalsWithTolerance(other.edge, tolerance) -> false
+            else -> true
+        }
+
+        fun transformBy(
+            transformation: Transformation,
+        ): Link = Link(
+            start = start.transformBy(transformation = transformation),
+            edge = edge.transformBy(transformation = transformation),
+        )
+    }
+
+    val links: List<Link>
 
     val segmentCurves: List<PrimitiveCurve>
 }
