+++ app/src/jvmMain/kotlin/diy/lingerie/geometry/closedSplineSvgDumping.kt
@@ -8,7 +8,7 @@ import diy.lingerie.simple_dom.svg.SvgPath
 fun ClosedSpline<*>.toSvgPath(
     stroke: SvgPath.Stroke = SvgPath.Stroke.default,
 ): SvgPath {
-    val edgeCurves = this.cyclicEdgeCurves
+    val edgeCurves = this.cyclicCurves
     val start = edgeCurves.first().start
 
     return SvgPath(
