+++ app/src/jvmTest/kotlin/diy/lingerie/geometry/ClosedSplineSvgTests.kt
@@ -2,7 +2,7 @@ package diy.lingerie.geometry
 
 import diy.lingerie.geometry.curves.bezier.MonoBezierCurve
 import diy.lingerie.geometry.splines.ClosedSpline
-import diy.lingerie.geometry.splines.SplineLink
+import diy.lingerie.geometry.splines.Spline
 import diy.lingerie.simple_dom.pt
 import diy.lingerie.simple_dom.svg.SvgPath
 import diy.lingerie.simple_dom.svg.SvgRoot
@@ -23,32 +23,32 @@ class ClosedSplineSvgTests {
         assertEqualsWithTolerance(
             expected = ClosedSpline.positionallyContinuous(
                 links = listOf(
-                    SplineLink(
+                    Spline.Link(
                         start = Point(32.551998138427734, 125.20800018310547),
                         edge = MonoBezierCurve.Edge(
                             firstControl = Point(43.84199905395508, 108.19000244140625),
                             secondControl = Point(50.24800109863281, 84.95600128173828),
                         ),
                     ),
-                    SplineLink(
+                    Spline.Link(
                         start = Point(65.63800048828125, 72.54399871826172),
                         edge = MonoBezierCurve.Edge(
                             firstControl = Point(82.27100372314453, 59.12900161743164),
                             secondControl = Point(108.29000091552734, 56.98699951171875),
                         ),
                     ),
-                    SplineLink(
+                    Spline.Link(
                         start = Point(131.3769989013672, 50.821998596191406),
                         edge = MonoBezierCurve.Edge(
                             firstControl = Point(126.60900115966797, 85.4229965209961),
                             secondControl = Point(146.66000366210938, 103.48999786376953),
                         ),
                     ),
-                    SplineLink(
+                    Spline.Link(
                         start = Point(181.7220001220703, 111.55500030517578),
                         edge = LineSegment.Edge,
                     ),
-                    SplineLink(
+                    Spline.Link(
                         start = Point(131.70599365234375, 177.86399841308594),
                         edge = MonoBezierCurve.Edge(
                             firstControl = Point(85.72000122070312, 174.21600341796875),
@@ -65,32 +65,32 @@ class ClosedSplineSvgTests {
     fun toSvgPathElementTest() {
         val closedSpline = ClosedSpline.positionallyContinuous(
             links = listOf(
-                SplineLink(
+                Spline.Link(
                     start = Point(32.551998138427734, 125.20800018310547),
                     edge = MonoBezierCurve.Edge(
                         firstControl = Point(43.84199905395508, 108.19000244140625),
                         secondControl = Point(50.24800109863281, 84.95600128173828),
                     ),
                 ),
-                SplineLink(
+                Spline.Link(
                     start = Point(65.63800048828125, 72.54399871826172),
                     edge = MonoBezierCurve.Edge(
                         firstControl = Point(82.27100372314453, 59.12900161743164),
                         secondControl = Point(108.29000091552734, 56.98699951171875),
                     ),
                 ),
-                SplineLink(
+                Spline.Link(
                     start = Point(131.3769989013672, 50.821998596191406),
                     edge = MonoBezierCurve.Edge(
                         firstControl = Point(126.60900115966797, 85.4229965209961),
                         secondControl = Point(146.66000366210938, 103.48999786376953),
                     ),
                 ),
-                SplineLink(
+                Spline.Link(
                     start = Point(181.7220001220703, 111.55500030517578),
                     edge = LineSegment.Edge,
                 ),
-                SplineLink(
+                Spline.Link(
                     start = Point(131.70599365234375, 177.86399841308594),
                     edge = MonoBezierCurve.Edge(
                         firstControl = Point(85.72000122070312, 174.21600341796875),
